+++ src/seppl/io/__init__.py
@@ -1,6 +1,6 @@
 from ._reader import Reader, DirectReader
-from ._writer import Writer, DirectWriter, StreamWriter, DirectStreamWriter, BatchWriter, DirectBatchWriter
-from ._filter import Filter, MultiFilter, FILTER_ACTIONS, FILTER_ACTION_KEEP, FILTER_ACTION_DISCARD
+from ._writer import Writer, DirectWriter, StreamWriter, DirectStreamWriter, BatchWriter, DirectBatchWriter, DataCollector
+from ._filter import Filter, StreamFilter, FILTER_ACTIONS, FILTER_ACTION_KEEP, FILTER_ACTION_DISCARD, filter_data, FilterPipelineIterator, MultiFilter
 from ._execution import execute
 from ._split import gcd, Splitter, add_splitting_params, init_splitting_params, transfer_splitting_params, initialize_splitting
 from ._utils import locate_files